chore: move evm cli-utils to common
