+++ evm/gateway.js
@@ -24,7 +24,7 @@ const {
     httpGet,
     getContractJSON,
 } = require('./utils');
-const { addBaseOptions } = require('./cli-utils');
+const { addBaseOptions } = require('../common');
 const { getWallet } = require('./sign-utils');
 
 const IGateway = require('@axelar-network/axelar-gmp-sdk-solidity/interfaces/IAxelarGateway.json');
